@@ -1,6 +1,6 @@
 module github.com/rorski/templizer
 
-go 1.17
+go 1.18
 
 require sigs.k8s.io/yaml v1.3.0
 
