@@ -7,7 +7,7 @@ import (
 
 // HCLJoin converts a list of interfaces to a list of strings in HCL/Terraform friendly format
 // For example, []string{"this", "that", "the other"} wouid produce: ["this", "that", "the other"]
-func HCLJoin(values []interface{}) string {
+func HCLJoin(values []any) string {
 	var s []string
 	for _, v := range values {
 		v := fmt.Sprintf("\"%s\"", v.(string))
